Moved profile page to profile component, and called the component in the page
